gui: Make app icon creation folder independant
So far the path to the de1app sourcefolder had to be placed directly under /sdcard.
This leads to the de1app repository beeing not directly executable. While the app would run
the icon creation would not.
To fix that we are querring the current scripts folder to find the correct path from there.

Signed-off-by: Mimoja <git@mimoja.de>