set awake flag if we are mid-schedule at startup
This rolls back the last commit (adding the -nocomplain option) and fixes the problem properly. The issue is that at startup we might be in the middle of a scheduled period, and if so the awake flag should be set. 

This accomplishes two things: (1) it ensures the machine doesn't sleep until the schedule ends if we are mid-schedule at startup; and (2) it avoids the possibility that the awake flag isn't set when we get to the end of our scheduled time.