Add plugin system
This commit adds a plugin system (without the UI) to be used to
support more extensions to the app that might not be beneficital
to all users. It takes inspiration from the skin subsystem.

Every plugin has its settings (if available) loaded before sourcing.
The settingsfile is hardcoded to $plugin-settings.tdb. The version
translation of settings is within the plugin responsibilities.

Only the top most level of the plugin folder is searched and sourced.
Plugins can therefore store additional data in subfolders.
The use of namespaces in encuraged, but not enforced by the code.

Tested on win32/undroidwish.

This commit also contains an example plugin showing how to trace functions
state-chagnes and work with settings.

Signed-off-by: Mimoja <git@mimoja.de>