Print settings when selecting an extension
As we do not want to source and extension code without the
uses consent we can not display any dynamic variables.
But: We can use the additional space to show the saved settings
that are in use. This allows us to double-check the current configuration
before activating a plugin. We can not easily open the settings file
in an external android editor without the risk of exposing user data to
3rd party apps without the user consent. I personally dont think we can
solve that from within the app (for now). User will still need to edit their
plugin settings from within the filebrowser / adc / their app of choice.

Signed-off-by: Mimoja <git@mimoja.de>