plugins: Add plugin tabs
In this commit we are removing the scentone heart button from the Insight skin
(so far without replacement) and replace it with a plugin Tab overview.
Plugins can be attached to Tabs if they support this. To support this feature
they need to support a preload method that hand them the context to draw into.
Therefore an App restart is required between Tab changes.

We are also updating the psd submodule at once to include the new tab base files.

Signed-off-by: Mimoja <git@mimoja.de>