Introduce new event system, rework scale-weight handling
The existing event-generation system only provides callbacks
for major-state changes. This granularity is not sufficient
for triggering on, for example, changes between readying,
flow, and flow complete (but not yet in Idle).

This commit provides a general event framework that allows
scheduling of callbacks synchronously, or on the event or idle
queues. Events for major and minor mode changes of the DE1,
as well as scale-connect/disconnect are provided.

The current algorithms for weight and mass-flow estimation
have significant delay. New algorithms are introduced here.

The "low drink weight" bug is addressed by moving the call of
the history-save function to the new, after_flow_complete event.

Warnings about a scale "failure" or starting an SAW shot
without the scale reporting weight are surfaced through
::gui::notify::scale_event and can be modified as needed.

The scale period is estimated from the arrival times
rather than assuming a 10 SPS rate.  A tau of ~17 minutes
is used and the estimate is persisted by device ID.

The stop-on-weight (SAW) and auto-tare logic has been rewritten
to improve robustness and reduce the BLE traffic.

Estimation approaches that may be appropriate for high-vibration
environments are present. These can be enabled through extensions,
but require data for examination before direct incorporation.

Many new predicates are introduced for the DE1's state and
the scale to reduce dependencies on internal representations.

Though when modified and appropriate, initialization and functionality
has been moved out of in-line code and into an appropriate namespace,
this is not a complete refactor; much in-line code remains unmodified.

SIGNIFICANT END-USER CHANGES
============================

Drink weight is tracked after the flow stops. A new toast, "Enjoy!"
indicates that the recording is completed. Certain skins may "jump"
when this period ends as they have not been updated completely.

$::settings(seconds_after_espresso_stop_to_continue_weighing)
sets this duration. Its default of 8 seconds is probably too long.
It only needs to be long enough for the scale to respond to the
last few drops falling at the end of the shot.

The wording of the SAW-triggered toast has been changed.

SAW timing has changed, due to the faster filtering as well as
inclusion of filter delays and scale response estimates.
$::settings(stop_weight_before_seconds) can be set through
the GUI Settings screen with most skins. It likely needs to be
reduced significantly.

THEME / DEVELOPER NOTES
=======================

Highlights of Changes

Note the logging enhancements in a previous commit.

Themes might be able to set the logging threshold
based on the users preferences so that there are
"always" logs available, at least for high severity.

::register_state_change_handler

    has been deprecated. See de1_de1.tcl for events
    that might replace its use. The logs will indicate
    how existing calls have been rewritten with lambda

::de1::event::on_flow_change_add
::de1::event::after_flow_complete_add

    may be interesting as the first splits up the major states
    and the second is called after the flow stops with a delay of
    $::settings(seconds_after_espresso_stop_to_continue_weighing)
    (but not before entering Idle).

::scale_tare

    has been deprecated. See ::device::scale::tare

$::device::scale::autotare_states

    contains an overidable list of states for which
    auto-tare is active.

$::de1(final_water_weight)
$::de1(final_espresso_weight)

    are now used for their respective modes,
    rather than overloading final_water_weight.

$::settings(drink_weight)

    continues to be available for the weight-tracking states of
    Espresso and HotWater.

::finalwaterweight_text

    no longer provides both Espresso and HotWater weights.

::drink_weight_text

    is introduced to provide a mode-independent replacement.

::gui::notify::scale_event

    is called rather than having "borg toast" called directly
    from code that is handing the operation of the scale.
    Overrides in messaging are now possible within the GUI.

.shot file

    now additionally includes raw (as-recieved) scale data at full
    time resolution with timestamps as well as a zero-time reference
    and the timer to be able to reconstruct weight data during the shot

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>