scale: Improve final drink_weight estimate
The final_*_weight and drink_weight should agree with what is on
the scale before the cup is removed. As it is sometimes used
to calculate EY%, it should be as accurate as possible.

These changes track [weight_estimate] (by default the scale weight)
during flow and switch to the maximum of [final_weight_estimate]
(by default the median over 11 samples, ~1 second) after flow
to address the following use cases:

During flow:

* High flow into the cup an cause the scale to overestimate
  due to the force of the drink landing in the cup

  - Tracking, rather than taking the peak, should mitigate

* Cup or platform is bumped, leading to a momentary "false" weight

  - Tracking, rather than taking the peak, should mitigate

After flow:

* Drops still falling into the cup can lead to overestimation

  - Use of median filtering over ~1 second should result in
    a final estimate that reflects the "settled" weight

* Cup or platform is bumped after flow is complete

  - Use of median filtering should reject brief bumps

* Cup is removed before recording is complete

  - Capturing maximum of median should "remember" the weight

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>