scale: Allow warn_if_scale_not_reporting without target
The previous implemtation checked if am SAW target was set.
By removing the additional check within the proc, it now
can be used outside of the SAW functioality to inform the user
or take other actions.

The check for [::device::scale::expecting_present] is retained,
and "gating" of the checks is retained on the existing calls.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>