de1: Add on_connect/disconnect events
In order to be able to manage and reflect the state of the DE1,
it is desirable to be able to perform actions when the DE1
connects or disconnects without having to modify existing code.

This commit adds on_connect and on_disconnect events for the DE1
and fires them in the locations found within bluetooth.tcl.
There may be other locations that were not identified.

The on_disconnect event is intentionally not fired during
close_all_ble_and_exit {}

No existing code has been refactored to use thise events.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>