scale: Expand event_dict; refactor to improve understandability
Subsequent work indicated that expanding the event_dict
with a fuller set of estimates and associated times
would simplify consumers' code and make it more robust.

	event_time		[expr {[clock milliseconds] / 1000.0}]
	reported_weight		$reported_weight
	reported_time		$event_time
	weight			$_weight
	weight_time		$_weight_time
	weight_filtered		$_weight_filtered
	weight_filtered_time	$_weight_filtered_time
	flow			$_flow
	flow_time	 	$_flow_time
	flow_filtered		$_flow_filtered
	flow_filtered_time	$_flow_filtered_time
	scale_is_recording	[::device::scale::history::is_recording]
	this_state		[::de1::state::current_state]
	this_substate		[::de1::state::current_substate]

(Current code is always the best reference)

Due to API enhancement, package version bumped to 1.1

Naming and logical flow were adopted from previous code.
In some cases, the older names were confusing about what
the variable or series contained ("raw" as an example).

This commit moves to a more descriptive and accurate names
and reorganizes flow to be clearer in sections that mimiced
the early implementations.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>