de1: Properly fire after_flow_complete event when deferred
The after_flow_complete was not firing when deferred, waiting for Idle
This resulted in failure to record .shot files, as well as triggering
end-of-shot actions for skins and extensions.

This was only occuring with "ending" substates longer than the delay
(default of 5 seconds). This would have logged in the following shot
"Pending after_flow_complete callbacks. Second flow started before %g seconds?"

DEVELOPER NOTES
===============

::de1::event::apply::after_flow_is_pending
::de1::event::apply::after_flow_cancel_pending

renamed to

::de1::event::apply::after_flow_complete_is_pending
::de1::event::apply::after_flow_complete_cancel_pending

for clarity and consistency

Version bump due to API change.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>