de1: Disable SAV for "basic" profiles if scale is configured
Beta testing revealed challenges with "basic" profiles stopping early
due to unrealistically low SAV levels and a cumbersome UX for changing
those levels at this time, requiring "deleting" the scale and multiple
app restarts.

Disabled in-app SAV for basic profiles:

  settings_2a -- basic pressure
  settings_2b -- basic flow

when a scale is expected to be present.

::de1::sav::skip_sav_check can be overriden by skins or extensions

Changed wording of "Volume based stop ..." log message
to be consistent with "Weight based stop ..."

Version bump due to API addition

SIGNIFICANT END-USER CHANGES
============================

When a scale is configured, SAV is disabled for "basic" profiles
much as it was on the Stable branch prior to 1.35

If the scale is not connected or not reporting weights, the user
will be notified about the issue with toasts (earlier commits)
but SAV will no longer "take over".

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>