logging: Log background errors
Timestamps or details of background errors were not
captured in the logs with the default bgerror handler.
Log the details of the error then defer to the
standard bgerror handler.

Users will see the modal popup that Tk provides
currently. They can continue to use it as before
keeping the current behavior, including "skip".

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>