plugins: Log errors due to loading plugin
If there is a plugin-load error, the one-line error message is often not
sufficient to be able to diagnose the problem, or supply information
to the developer.

Extend the current "catch" around the skin load to capture the options dict
and log the contents. Retain the current `info_page` and disable behavior.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>