build-info: Read build-info, log, and make available in ::app
The build-info file that can be generated in the build process
contains detailed information about the build that can uniquely
identofy the contents of the build both in machine-readable and
human-readable forms.

Read in the build-info file if it is readable, output to the log,
and read into the ::app::build_info dict for other uses.

Provide ::app::version_string and ::app_build_timestamp from
build-info, using version.tcl and timestamp.txt as fallbacks.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>