Remove "stale" profile and settings requests from BLE queue
Users have reported unresponsive BLE after interacting with profiles.
In one case, a queue depth of over 1300 was observed. Examination
of logs suggested that at least one cause was users updating profiles
and settings faster than they could be transferred to the DE1.

As the settings that may be pending in the BLE queue are about to be
overwrtten, remove "matching" entries prior to adding their replacements.

remove_matching_ble_queue_entries {comment_regexp} is used to match
against the "comment" that is provided to userdata_append

This is a "spot fix" and should be reexamined if the BLE queue
is rewritten in the future. The following calls are addressed:

  * de1_send_shot_frames
  * de1_send_steam_hotwater_setting
  * set_steam_flow
  * set_steam_highflow_start
  * set_tank_temperature_threshold

Additional calls may need to be addressed in a similar manner

NB: This manipulates the queue directly and is not thread-safe

    This is not deemed to be an issue at this time as the Tcl
    implementation is single-threaded.

    Should this change in the future, the queues will need mutex

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>