GUI: Allow skins to provide non-resized splash screen
So far skins do not offer a way to provide a custom splash screen.
This PR adds a simple check if the [skin_directory] already contains a
properly sized splash screen. It is not resized if it does not exist
in the proper resolution to not increase the startup time.

Signed-off-by: Mimoja <git@mimoja.de>