scale: Allow wider range of SAW offsets
Spouted portafilter handles can significantly increase the delay
needed to hit SAW eeights. One report suggested that an offset of
around 1.2 seconds would be needed.

Increase the positive limit to 1.5 seconds and adjust the "helper"
intended to ease transition from earlier versions where 2.8 seconds
was the default.

It is unlikely that the negative limit will need to be changed as
1.0 seconds is already enough to back out the delay compensation.
More negative would mean "after the scale reaches target" or that
the delay estimate for the weight and mass-flow algorithm
implemented for SAW is not correct.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>