Dont set remote_timestamp to 0 when no update is availabe
check_timestamp_for_app_update_available returned 0 when no update is available.
In the case the user hit start_app_update manually it will lead to the local
time stamp beeing set to that return value instead of the remote value.

This leads to the app always showing an update available if the last update
was manually tapped.
We now return the local_timestamp on success instead of 0 and dont set
the global "update_available" flag instead of returning 0.

Signed-off-by: Mimoja <git@mimoja.de>