logging: Log at INFO for Stable builds; plugin to set to DEBUG
Stable builds generally should not require DEBUG-level logging with
the default set to INFO. At INFO, omce the severity has been set on
the majority of `msg` calls, normal operation should be able to be
confirmed by examining the logs, reducing the load on support.

For the cases where abnormalities appear on Stable builds, the logs
can be increased to DEBUG level by enabling the log_debug plugin.
Although this does not cover the period between app start and the
load and start of plugins, it should be sufficient to provide
visibility into operational issues.

Logging to the console and Android logging are not impacted.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>