skins: settings: Change logging option to indicate "always on"
Previous commits enabled logging for all users and ignored ::settings(log_enabled)

This commit changes the checkbox on the "default" Settings > Misc
screen to disabled and to read "Logging is enabled".

This choice is intended to provide messaging at the point where a user,
following older instructions, might try to enable logging.

In the future, skins may wish to provide an option to Stable-branch
users between DEBUG and INFO levels. At this time, the Log DEBUG
plugin can be activated, when a Stable-branch user or the support team
needs to examine logs at the DEBUG level.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>