scale: Reduce to single watchdog for updates
With commit a3f259f6, background errors are now logged. As a result,
errors in scale processing can be detected from the logs without use
of a watchdog on both entry and leave of processing the update.

Remove the watchdog at the completion of processing. Adjust variables
and GUI event handlers for a single watchdog.

Moved ::device::scale::show_update_watchdog_notifications to
::gui::notify::show_scale_update_watchdog_notifications

Changed wording on update interruption to "Check scale"

Version bumps due to name and signature changes.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>