scale: Add "Scale reporting" scale event when updates start arriving
Long-running processes in other parts of the app can exceed the 1-second
timeout for scale updates. Previously, only a TIMEOUT message was
presented to the user by the GUI, leading to comfusion as there was no
indication that the scale was still connected and operational.

Add a new ::gui::notify::scale_event for scale_reporting and trigger
when the first update comes through. This is expected to appear once
after connection and initialization of the scale, as well as after a
watchdog timeout, once updates are processed again. Examination of the
logs suggest that in most cases, the scale-reporting indication will
happen immediately after the former TIIMEOUT indication, due to the
way the event queue seems to work. One log examined showed 3-5 ms.

The two messages in the default implementation are changed to be:

timeout_updates: "Check scale"
scale_reporting: "Scale reporting"

At least in English, the longer text of the scale_reporting message
should immediately obscure the timeout-updates message.

These can be changed by skins or extensions.

NB: Versions NOT updated here due to the bum in the previous commit
"scale: Reduce to single watchdog for updates"

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>