de1: Protect against unrecognized states and substates
User reported a condition where, when using a refill kit, after
starting the DE1 but before the tray had been filled, a substate of
decimal 20 was reported. As this is not in the arrays in machine.tcl,
an error occurred. As the current state and substate are often checked
on each screen refresh, the GUI became non-responsive.

Provide a mitigating fix by capturing and logging the lookup failure.
As the substate is not known to the app or skin code, this should not
impact any exisitng functionality.

suggest backport to stable.

Resolve further in the future with updates to machine.tcl for any
known-missing states and substates.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>