gui: Suppress scale disconnect messages when DE1 is in Sleep
User reported that every 30 seconds, evne when sleeping, the toast
"Warning: Scale not connected" is displayed, along with a ticking sound.

This can be replicated by configuring a scale, turning it off, and setting
automatically_ble_reconnect_forever_to_scale 1

It is due to the `ble` stack sending a connection event as it starts
to try to connenct by address and then a disconnect event when it fails.

Gate the message so it is not displayed when the DE1 is in Sleep
state.

Long-term resolution is a BLE stack that properly reports
connection state.

Resolves https://github.com/decentespresso/de1app/issues/129

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>