logging: Add utilities for removing binary from logs
Many existing messages contain "raw binary" and, as such, are not
readable in the logs. Presence of sume of these characters can prevent
copy-paste from the logs as well as being interpreted as display
commands if rendered on the screen.

This commit introduces a set of "sanitizing" routines for
generic strings and lists, MMR, and ble data.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>