logging: Remove binary logging from bluetooth.tcl and de1_comms.tcl
Binary data in the logs makes them difficult to read or to copy-paste
for diagnosis. The Bluetooth communication previously logged "raw"
binary data in many places.

Adjust the logging in bluetooth.tcl and de1_comms.tcl to use the new
utilities in ::logging to "sanitize" the logs by converting them to a
printable format. As the communications are split between these two
files, each uses a "custom" logger to indicate the file from which the
message came. Consstency was improved, with a Bleutooth transaction
being logged with:

  ENQ	Enqueued (added to the pending Bluetooth command list)
  DEQ	Dequeued (sent to the `ble` module)
  ACK	Acknowledgement of the request recieved

Association of ACKs with DEQs is not done at this time.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>