Check still on same page before unhiding widgets after 400 ms
This solves the bug introduced by DUI when fast_tap_mode=0, widgets
disabled initially and shown after 400 ms to capture phantom tags could
appear on another page if the user changed pages very fast.
Now it is checked that we're still on the page before showing.
Also, now the widgets are not set to normal state after 400 ms if their
initial state is "disabled".