de1: Pop error page if reported state is FatalError or unable to decode
There have been reports of "20" and "-40" in bgError boxes. Neither of
these are present in the API. The "-40" is likely an error in decoding
216 (Error_BootFill).

This commit replaces the use of ::fields for decoding the StateInfo
payload.

Should additional or otherwise unexpected states or substates arise,
also catch any decode errors.

If FatalError is reported as the DE1 state, or there is an error
decoding the state and substate, raise a message using the info_page
if available, with a toast if that fails.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>