de1: Provide emergency shutdown and trigger on error StateInfo
If the DE1 reports an error condition, the firmware will likely have
already put it into "self-protect mode". As a back-up to this, disable
all queued BLE commands and the queueing of further commands.

Immediately send a RequestState Idle and another in 500 ms.

The single, pending queue of BLE commands is also cleared.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>