gui: Provide meaningful error messages on FatalError
Using info_page, provide an error message to the user with information
for the support team, as well as if there is likely a resolution that
they can persue themselves.

In the case of Error_BootFill, clear instructions about filling the
tank, even with a catering or plumbing kit are given.

For Error_HiCurrent and Error_LoCurrent, turning off the DE1 is
instructed, rather than a reboot with other error codes.

If info_page is not available, the beError box should contain at least
the "headline", including the state and substate, in readable form.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>