de1: Add volume_dispensed to on_shotvalue_available
Although the flow rate and timing is already available in the
event_dict for on_shotvalue_available, this is a common-enough
value needed by the GUI and shot-control extensions to have it
available in the event_dict directly.

Added "volume_dispensed" to the event_dict as the total of

  $::de1(preinfusion_volume)
  $::de1(pour_volume)

Frame information is available to consumers in the same event_dict
so that division by flow-phase, frame, or time, is possible.

Package version bump due to additive API change.

Signed-Off-By: Jeff Kletsky <git-commits@allycomm.com>