scale, bookoo: Cleanup the code and add battery and flow parsing
The BooKoo Scale does indeed report its battery level as well as a flow.
While there is no place to plug in scale flow (for now) as an additional
datapoint, we can prepare it by parsing and use the battery level as well
for the app analog to the felicita and decent scale.

Signed-off-by: Mimoja <git@mimoja.de>