Improve starting estimate for DiFluid and avoid cross-scale contamination of period estimation (#333)
* Avoid cross-scale period estimation contamination

When a new scale is connected after an old scale was connected,
the estimate state was not reset. Move the initialization into
the init procedure such that the new scale starts with a clean
slate.

* Improve starting estimates for DiFluid scale

Adjust the starting estimate for difluid to be the nomimal
reporting period of 200ms, as confirmed by DiFluid.

Given my scale reporting time of about 225ms, this
results in convergence to various error levels at:

        |       error
time    |        10%   |      1%
--------|----------------------------
before  |    ~64 mins  | ~150 mins
after   |    ~4 mins   |  ~90 mins

In all likelihood, we hence achieve usable results by 1st
or second shot.

Fixes #331 (or well significantly improves it :D)